Added to "The Scientist"
Lyra convinces Anon to let her sketch his nude body. Anon, after getting
comfortable with the idea, gets less and less so when she begins to
focus on his genitals, violating him to give him an erection to sketch.
When Lyra demands to see him orgasm, Anon tries to leave, only for Lyra
to bind him to the chair.