Added to "The Scientist"
Anon wakes up from his nightmare, fueled by the memories of his
captivity, a month or so after escaping. After spending a bit in the
hospital, he's now in a wheelchair and has a cast on his arm, barely
able to sleep, and paranoid about Lyra, despite knowing that the guard
killed her while rescuing him. He goes about his morning routine, nearly
flips out when he hears the creaky stairs of the apartment building, and
checks the anti-teleportation talismans he bought, before locking the
multitude of locks on his front and bedroom door and lying down for a
nap.